--- a/hashable.cabal
+++ b/hashable.cabal
@@ -39,6 +39,7 @@ Flag sse41
 
 Library
   Exposed-modules:   Data.Hashable
+                     Data.Hashable.Lifted
   Other-modules:     Data.Hashable.Class
   Build-depends:     base >= 4.0 && < 4.10,
                      bytestring >= 0.9 && < 0.11