… Data/Hashable/Lifted.hs
… benchmarks/Benchmarks.hs
… examples/Main.hs
… tests/Properties.hs