This implements what is proposed in https://github.com/tibbe/hashable/issues/114. It is not yet complete. I still need to add more instances. There are two deviations from the original proposal:

1. The method `liftHash` is not included. The reason for this are found in the documentation in `Data.Hashable.Lifted`.
2. The `Hashable2` class was included as well. This is mostly for consistency with other packages that provide higher rank typeclass variants.

In some of existing `Hashable` instances, I call `hashWithSalt1` to avoid duplicating code. You can look at the instance for `[]` to see this. I do not believe that this will hurt performance because the inliner should expand these.

The example instances I provide in `Data.Hashable.Lifted` have not been typechecked and are likely broken. I'll get to that soon.

Let me know at this point if there are any concerns so that the can be addressed before I write all the remaining instances.