@RyanGlScott 2016-10-19
Generics support for `Hashable1`?