@andrewthad 2016-10-19
@RyanGlScott I knew you would show up ;) Now that I think about it, that would be useful (especially if I'm working with `Fix Expr`, and `Expr` and I keep changing `Expr`). I'll attempt to implement this, and I would appreciate if you would review it.