@RyanGlScott 2016-10-19
> I'll attempt to implement this, and I would appreciate if you would review it.

Sure thing! I recommend using a trick like `aeson`'s [`ToArgs`](https://github.com/bos/aeson/blob/8044d1a54d4dc619a9e740c4c651795807b0abc0/Data/Aeson/Types/ToJSON.hs#L164-L166) as a way to reuse the code between the code for the `Generic`- and `Generic1`-based machinery.