@andrewthad 2016-10-19
@RyanGlScott Thanks. That's a really good trick. I was just about to duplicate everything.