@andrewthad 2016-10-19
Generics for `Hashable1` is now ready. Also, does anyone know why `hashSum` even takes `size` as a parameter? It looks like there some bitshifting arithmetic that happens on it, but ultimately, it's never used. Here it is as it currently exists: https://github.com/tibbe/hashable/blob/a5bb09f8ecd175e62118258ffd30ac4afaf35530/Data/Hashable/Generic.hs#L43-L61

I'm guessing that it's leftover from a different approach to hashing sums that was removed.