@RyanGlScott 2016-10-19
@andrewthad I think you're missing a `GHashable` case for `(:.:)` (i.e., how would you generically handle something like `newtype Compose f g a = Compose (f (g a))`?)