@andrewthad 2016-10-19
@RyanGlScott Thanks for the feedback. I've address both of your comments.