@andrewthad 2016-10-21
Not sure what's causing the build failure. The failing line is:

    Preprocessing test suite 'tests' for hashable-1.2.4.0...
<command line>: cannot satisfy -package-id hashable-1.2.4.0-inplace: 
    hashable-1.2.4.0-inplace is shadowed by package hashable-1.2.4.0-85f768408d56687a048b7dee7b04a672

Other than that, I still need to add tuple instances. Any other thoughts or concerns?