@andrewthad 2016-10-21
Thanks for clearing that up. I've wondered about test suites that use other-modules like that before. Definitely won't try to fix that in this PR.

I'll go ahead and get the tuple instances ready.