@andrewthad 2016-10-22
Ben Gamari was able to help look into the differences in the generated core, and the new generic implementation of `hashWithSalt` surprisingly results in fewer allocations. I've added benchmarks the generic machinery. Here are the results:

    ---------------------------
    NEW GENERICS IMPLEMENTATION
    ---------------------------
    > cabal bench --benchmark-options='Generic'

    Preprocessing benchmark 'benchmarks' for hashable-1.2.4.0...
    Running 1 benchmarks...
    Benchmark benchmarks: RUNNING...
    benchmarking Generic/product
    time                 25.82 ns   (25.77 ns .. 25.88 ns)
                         1.000 R²   (1.000 R² .. 1.000 R²)
    mean                 25.85 ns   (25.80 ns .. 25.98 ns)
    std dev              257.0 ps   (118.7 ps .. 477.9 ps)

    benchmarking Generic/sum
    time                 5.756 ns   (5.749 ns .. 5.765 ns)
                         1.000 R²   (1.000 R² .. 1.000 R²)
    mean                 5.757 ns   (5.751 ns .. 5.768 ns)
    std dev              24.81 ps   (14.63 ps .. 42.26 ps)

    benchmarking Generic/product and sum
    time                 8.425 ns   (8.415 ns .. 8.435 ns)
                         1.000 R²   (1.000 R² .. 1.000 R²)
    mean                 8.422 ns   (8.415 ns .. 8.432 ns)
    std dev              27.64 ps   (20.11 ps .. 36.48 ps)

    Benchmark benchmarks: FINISH

    ---------------------------
    OLD GENERICS IMPLEMENTATION
    ---------------------------
    > cabal bench --benchmark-options='Generic'

    Preprocessing benchmark 'benchmarks' for hashable-1.2.4.0...
    Running 1 benchmarks...
    Benchmark benchmarks: RUNNING...
    benchmarking Generic/product
    time                 26.40 ns   (26.38 ns .. 26.44 ns)
                         1.000 R²   (1.000 R² .. 1.000 R²)
    mean                 26.42 ns   (26.39 ns .. 26.48 ns)
    std dev              128.1 ps   (91.03 ps .. 172.5 ps)

    benchmarking Generic/sum
    time                 9.844 ns   (9.832 ns .. 9.859 ns)
                         1.000 R²   (1.000 R² .. 1.000 R²)
    mean                 9.841 ns   (9.832 ns .. 9.857 ns)
    std dev              39.71 ps   (25.60 ps .. 64.43 ps)

    benchmarking Generic/product and sum
    time                 13.90 ns   (13.87 ns .. 13.93 ns)
                         1.000 R²   (1.000 R² .. 1.000 R²)
    mean                 13.91 ns   (13.89 ns .. 13.93 ns)
    std dev              70.92 ps   (53.18 ps .. 99.33 ps)