@andrewthad 2016-10-24
Also, just in case anyone wants to look at master but with the backported benchmarks, here it is: https://github.com/andrewthad/hashable/tree/master_with_examples. I think that I still need to correct some of the examples I gave in the haddocks.

@tibbe What were your thoughts on my [response concerning Hashed](https://github.com/tibbe/hashable/pull/124#discussion_r84583839)?